requirements.txt
backcall==0.1.0
Click==7.0
decorator==4.3.0
Flask==1.0.2
Flask-HTTPAuth==3.2.4
ipython==7.0.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.13.1
Jinja2==2.10
MarkupSafe==1.0
parso==0.3.1
pexpect==4.6.0
picamera==1.13
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==2.0.5
ptyprocess==0.6.0
Pygments==2.2.0
RPi.GPIO==0.6.3
simplegeneric==0.8.1
six==1.11.0
timeout-decorator==0.4.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
